浮點單數據類型從float變更為double

+++ twant/app/src/main/java/com/ftofs/twant/adapter/ConfirmOrderStoreAdapter.java
@@ -88,13 +88,13 @@ public class ConfirmOrderStoreAdapter extends BaseMultiItemQuickAdapter<MultiIte
             }
 
             //每家商店小計金額
-            float realFreightAmount = item.freightAmount;
+            double realFreightAmount = item.freightAmount;
             //
             if (payWayIndex == 1) {  // 如果是門店自提，則不算運費
                 realFreightAmount = 0;
             }
             // float finalPayAmount = item.buyItemAmount + realFreightAmount - item.discountAmount;
-            float taxAmount =item.taxAmount;
+            double taxAmount =item.taxAmount;
             SLog.info(item.toString());
             double finalPayAmount = item.buyItemAmount + realFreightAmount;
             if (item.tariffEnable == Constant.TRUE_INT) {
