+++ twant/app/src/main/java/com/ftofs/twant/entity/ConfirmOrderStoreItem.java
@@ -53,9 +53,9 @@ public class ConfirmOrderStoreItem implements MultiItemEntity {
     public int storeId;
     public String storeName;
     public double taxAmount=0;//稅費
-    public float buyItemAmount;  // 金額
-    public float freightAmount; // 運費
-    public float discountAmount;  // 商店優惠
+    public double buyItemAmount;  // 金額
+    public double freightAmount; // 運費
+    public double discountAmount;  // 商店優惠
     public int itemCount;    // 商店訂單的產品件數: 如果sku1有2件，sku2有3件，那么件數就是5
     public String leaveMessage;  // 留言
     public int voucherCount;  // 商店券數量
